3822 - Added architecture detection for Windows 32 and 64 bit.  Fixed powershell scripts so that they would run on powershell v2. Added pcap replay functionality to sensors. Fixed ruleset sync. Fixed DNS resolution issues.

+++ web/backend/app/service/rulesync_service.py
@@ -4,6 +4,7 @@ import json
 import logging
 import os
 import sys
+import traceback
 
 from app.service.socket_service import NotificationMessage, NotificationCode, notify_ruleset_refresh
 from shared.connection_mngs import (FabricConnection, MongoConnectionManager,
@@ -14,25 +15,12 @@ from fabfiles import (KIT_ID, decode_password, KICKSTART_ID,
                       RULESET_STATES, NODE_TYPES, RULE_TYPES)
 from fabfiles.utils import get_suricata_pod_name, get_bro_pod_name
 from logging.handlers import RotatingFileHandler
-from logging import Logger, StreamHandler
 from pymongo.results import UpdateResult
 from typing import List, Dict
 
 
-LOG_FILENAME = "/var/log/tfplenum/rulesync.log"
 SURICATA_RULESET_LOC = "/opt/tfplenum/suricata/rules/suricata.rules"
-BRO_CUSTOM_DIR = "/opt/tfplenum/bro/scripts/custom"
-logger = logging.getLogger('rulesync_logger')
-
-
-def _setup_logger(log_handle: Logger, max_bytes: int=10000000, backup_count: int=3):
-    handler = RotatingFileHandler(LOG_FILENAME, maxBytes=max_bytes, backupCount=backup_count)
-    log_handle.setLevel(logging.DEBUG)
-    formatter = logging.Formatter('%(levelname)7s:%(asctime)s:%(filename)20s:%(funcName)20s():%(lineno)5s:%(message)s')
-    # sysout = StreamHandler(sys.stdout)
-    handler.setFormatter(formatter)
-    log_handle.addHandler(handler)
-    # log_handle.addHandler(sysout)
+BRO_CUSTOM_DIR = "/opt/tfplenum/zeek/scripts/"
 
 
 class RuleSyncError(Exception):
@@ -46,6 +34,12 @@ class RuleSynchronization():
         self.rule_sets = None # type: List[Dict]
         self.mongo = None # type: MongoConnectionManager
         self.notification = NotificationMessage(role="rulesync")
+
+    def _send_notification(self, msg: str, code: str = NotificationCode.ERROR.name):
+        print(msg)
+        self.notification.setStatus(status=code)
+        self.notification.setMessage(msg)
+        self.notification.post_to_websocket_api()
 
     def _is_sensor_in_ruleset(self, ip_address: str, rule_set: Dict) -> bool:
         if rule_set["sensors"] is None or len(rule_set["sensors"]) == 0:
@@ -71,12 +65,12 @@ class RuleSynchronization():
                     if rule_set["isEnabled"]:
                         ids_to_reset.append(rule_set['_id'])
                 except KeyError as e:
-                    logger.error("Failed to clear " + str(rule_set))
+                    self._send_notification("Failed to clear " + str(rule_set))
 
         self.mongo.mongo_ruleset.update_many({'_id': {'$in': ids_to_reset}}, {'$set': {'state': []}})
 
     def _build_suricata_rule_file(self, ip_address: str) -> io.StringIO:
-        logger.info("Building suricata rule file for {}.".format(ip_address))
+        print("Building suricata rule file for {}.".format(ip_address))
         rules_file = io.StringIO()
         for rule_set in self.rule_sets:
             if rule_set["appType"] != RULE_TYPES[0]:
@@ -93,7 +87,7 @@ class RuleSynchronization():
                     if rule["isEnabled"]:
                         rules_file.write(rule["rule"] + "\n")
                 except KeyError as e:
-                    logger.warning("Failed to write suricata rule {} because it is missing {} field inside of the object.".format(rule, str(e)))
+                    self._send_notification("Failed to write suricata rule {} because it is missing {} field inside of the object.".format(rule, str(e)))
 
         rules_file.seek(0, os.SEEK_END)
         if rules_file.tell() == 0:
@@ -114,7 +108,7 @@ class RuleSynchronization():
                     state_obj = {"hostname": hostname, "state": statestr}
                     ret_val = self.mongo.mongo_ruleset.update_one({"_id": rule_set["_id"]}, {"$push": {"state": state_obj}}) # type: UpdateResult
                     if ret_val.modified_count == 0:
-                        logger.error("Failed to update rule set ID {} with {}".format(rule_set["_id"], str(state_obj)))
+                        self._send_notification("Failed to update rule set ID {} with {}".format(rule_set["_id"], str(state_obj)))
 
     def _set_suricata_states(self,
                     hostname: str,
@@ -139,7 +133,13 @@ class RuleSynchronization():
         raise ValueError("Failed to find Suricata pod name.")
 
     def _sync_suricata_rulesets(self, fabric: Connection, ip_address: str, hostname: str):
-        suricata_pod_name = get_suricata_pod_name(ip_address, self.mongo)
+        try:
+            suricata_pod_name = get_suricata_pod_name(ip_address, self.mongo)
+        except ValueError as e:
+            print(str(e))
+            self._send_notification("Suricata is not installed on {}".format(hostname), NotificationCode.CANCELLED.name)
+            return
+
         rules_file = self._build_suricata_rule_file(ip_address)
         reload_cmd = 'kubectl exec -t {} -- /suricata/bin/suricatasc -c reload-rules'.format(suricata_pod_name)
         fabric.put(rules_file, SURICATA_RULESET_LOC)
@@ -147,10 +147,10 @@ class RuleSynchronization():
         message = json.loads(ret_val.stdout)
         if (ret_val.return_code == 0 and message['message'] == 'done' and message['return'] == 'OK'):
             self._set_suricata_states(hostname, ip_address, RULESET_STATES[2])
-            logger.info("Successfully synchronized Suricata signatures for {}.".format(hostname))
+            self._send_notification("Successfully synchronized Suricata signatures for {}.".format(hostname), NotificationCode.DEPLOYED.name)
         else:
             self._set_suricata_states(hostname, ip_address, RULESET_STATES[3])
-            logger.error("Failed to synchronize Suricata signatures on pod {} on node {}".format(suricata_pod_name, ip_address))
+            self._send_notification("Failed to synchronize Suricata signatures on pod {} on node {}".format(suricata_pod_name, ip_address))
 
     def _run_cmd_or_raise(self, fabric: Connection, cmd: str):
         ret_val = fabric.run(cmd)
@@ -167,7 +167,13 @@ class RuleSynchronization():
         return name.replace(' ', '_')
 
     def _sync_bro_rulesets(self, fabric: Connection, ip_address: str, hostname: str):
-        bro_pod_name = get_bro_pod_name(ip_address, self.mongo)
+        try:
+            bro_pod_name = get_bro_pod_name(ip_address, self.mongo)
+        except ValueError as e:
+            print(str(e))
+            self._send_notification("Zeek is not installed on {}".format(hostname), NotificationCode.CANCELLED.name)
+            return
+
         bro_load_file = io.StringIO()
         self._run_cmd_or_raise(fabric, "rm -rf %s/*" % BRO_CUSTOM_DIR)
 
@@ -193,25 +199,27 @@ class RuleSynchronization():
                     bro_file = io.StringIO()
                     bro_file.write(rule['rule'])
                     bro_file.seek(0, os.SEEK_END)
-                    logger.debug(bro_file.getvalue())
                     rule_name = self._convert_bro_rule_name(rule['ruleName'])
                     bro_file_path = "%s/%s" % (rule_set_dir, rule_name)
                     fabric.put(bro_file, bro_file_path)
                     self._run_cmd_or_raise(fabric, "chmod 0744 %s" % bro_file_path)
-
                     # Append the load statement
                     pos = rule_name.rfind('.')
                     bro_load_file.write('@load custom/%s/%s\n' % (rule_set_name, rule_name[0:pos]))
                 except KeyError as e:
-                    logger.warning("Failed to write bro script {} because it is missing {} field inside of the object.".format(rule, str(e)))
+                    self._send_notification("Failed to write Zeek script {} because it is missing {} field inside of the object.".format(rule, str(e)))
 
         bro_load_file.seek(0, os.SEEK_END)
+        if len(bro_load_file.getvalue()) == 0:
+            self._send_notification("Nothing to sync for bro rules.", NotificationCode.CANCELLED.name)
+            return
+
         bro_load_file_path = "%s/__load__.bro" % BRO_CUSTOM_DIR
         fabric.put(bro_load_file, bro_load_file_path)
         self._run_cmd_or_raise(fabric, "chmod 0744 %s" % bro_load_file_path)
         self._run_cmd_or_raise(fabric, "kubectl delete pod %s" % bro_pod_name)
         self._set_bro_states(hostname, ip_address, RULESET_STATES[2])
-        logger.info("Successfully synchronized Bro scripts for {}.".format(hostname))
+        self._send_notification("Successfully synchronized Zeek scripts for {}.".format(hostname), NotificationCode.DEPLOYED.name)
 
     def sync_rulesets(self):
         try:
@@ -239,7 +247,7 @@ class RuleSynchronization():
                     hostname = node['deviceFacts']['hostname']
                     with FabricConnection(ip_address, password) as fabric:
                         try:
-                            logger.info("Synchronizing Suricata signatures for {}.".format(hostname))
+                            print("Synchronizing Suricata signatures for {}.".format(hostname))
                             self._sync_suricata_rulesets(fabric, ip_address, hostname)
                         except Exception as e:
                             self._set_suricata_states(hostname, ip_address, RULESET_STATES[3])
@@ -247,10 +255,10 @@ class RuleSynchronization():
                             self.notification.setMessage("Failed to synchronize Suricata rules for {}.".format(hostname))
                             self.notification.setException(e)
                             self.notification.post_to_websocket_api()
-                            logger.exception(e)
+                            traceback.print_exc()
 
                         try:
-                            logger.info("Synchronizing Bro scripts for {}.".format(hostname))
+                            print("Synchronizing Bro scripts for {}.".format(hostname))
                             self._sync_bro_rulesets(fabric, ip_address, hostname)
                         except Exception as e:
                             self.notification.setStatus(status=NotificationCode.ERROR.name)
@@ -258,9 +266,9 @@ class RuleSynchronization():
                             self.notification.setException(e)
                             self.notification.post_to_websocket_api()
                             self._set_bro_states(hostname, ip_address, RULESET_STATES[3])
-                            logger.exception(e)
+                            traceback.print_exc()
 
-            logger.info("Synchronization Complete!")
+            print("Synchronization Complete!")
             self.notification.setStatus(status=NotificationCode.COMPLETED.name)
             self.notification.setMessage("Rule synchronization complete successfully!")
             self.notification.post_to_websocket_api()
@@ -269,12 +277,11 @@ class RuleSynchronization():
             self.notification.setMessage("Unrecoverable error. This is really bad contact the programmers!")
             self.notification.setException(e)
             self.notification.post_to_websocket_api()
-            logger.exception(e)
+            traceback.print_exc()
 
 
 @celery.task
 def perform_rulesync():
-    _setup_logger(logger)
     rs = RuleSynchronization()
     rs.sync_rulesets()
     notify_ruleset_refresh()
